@@ -20,40 +20,33 @@
   },
   "private": true,
   "peerDependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
     "core-js": "^2.4.1",
     "moment": "^2.19.3",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.14",
     "font-awesome": "4.7.0",
     "getbase": "^3.5.1",
     "breakpoint-sass": "2.7.0"
   },
   "devDependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "@angular/cli": "1.5.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/language-service": "^5.0.0",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/animations": "^7.0.3",
+    "@angular/cli": "^7.0.5",
+    "@angular/common": "^7.0.3",
+    "@angular/compiler": "^7.0.3",
+    "@angular/compiler-cli": "^7.0.3",
+    "@angular/core": "^7.0.3",
+    "@angular/forms": "^7.0.3",
+    "@angular/http": "^7.0.3",
+    "@angular/language-service": "^7.0.3",
+    "@angular/platform-browser": "^7.0.3",
+    "@angular/platform-browser-dynamic": "^7.0.3",
+    "@angular/router": "^7.0.3",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "~3.2.0",
+    "breakpoint-sass": "2.7.0",
+    "core-js": "^2.4.1",
+    "font-awesome": "4.7.0",
+    "getbase": "^3.5.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
@@ -62,16 +55,12 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
+    "moment": "^2.19.3",
     "protractor": "~5.1.2",
+    "rxjs": "^6.3.3",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.4.2",
-    "rxjs": "^5.5.2",
-    "core-js": "^2.4.1",
-    "zone.js": "^0.8.14",
-    "moment": "^2.19.3",
-    "font-awesome": "4.7.0",
-    "getbase": "^3.5.1",
-    "breakpoint-sass": "2.7.0"
+    "typescript": "~3.1.6",
+    "zone.js": "^0.8.26"
   }
 }
