@@ -1,11 +1,10 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/observable/empty';
+import { Observable, of } from 'rxjs';
 
 @Injectable()
 export class UserprofileService {
 
     public getUserprofile(): Observable<any> {
-        return Observable.of(null);
+        return of(null);
     }
 }
