@@ -1,5 +1,5 @@
 import { Directive, OnInit, HostListener } from '@angular/core';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
+import { ReplaySubject } from 'rxjs';
 import { ScrollSpyService } from '../services/scrollSpyService';
 import { ScrollService } from '../services/scrollService';
 
