@@ -1,4 +1,4 @@
-import { ReplaySubject } from 'rxjs/ReplaySubject';
+import { ReplaySubject } from 'rxjs';
 
 export class GenericCache<T> {
     [name: string]: ReplaySubject<T>;
