@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { ReplaySubject } from 'rxjs/ReplaySubject';
+import { ReplaySubject } from 'rxjs';
 import { LoadingIndicatorEvent } from '../model/loadingIndicatorEvent';
 
 @Injectable()
